create chat when uploading the file for the first time
