+++ apps/browser/components/providers/Providers.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { Provider as JotaiProvider } from "jotai";
-import WorkspaceFilesProvider from "@/components/providers/WorkspaceFilesProvider";
 import ReactQueryProvider from "./ReactQueryProvider";
 import { SessionProvider } from "next-auth/react";
 import ToastProvider from "./ToastProvider";
@@ -12,7 +11,7 @@ export function Providers({ children }: { children: React.ReactNode }) {
       <JotaiProvider>
         <ToastProvider>
           <ReactQueryProvider>
-            <WorkspaceFilesProvider>{children}</WorkspaceFilesProvider>
+            {children}
           </ReactQueryProvider>
         </ToastProvider>
       </JotaiProvider>
